pyproject.toml
[tool.poetry]
name = "perplexity-api-client"
version = "0.1.0"
description = ""
authors = ["SkorpionG2000 <105552971+SkorpionG@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
